Annotated Changes
In ADoc book sources, mark all my (previous) changes to the original
text via comments starting in `// @EDITED`, and followed by the status
of the edit (i.e. whether documented or not, and whether approved or
still waiting for KT approval).

Previous edits already contained comments about the nature of the
changes and how the text was before and after. This commits just
introduced a formal system of annotation via "comment tags".