Footnotes: App B
Proofread and externalize all footnotes in "App.x B: THE HUGO LIBRARY".

There are two footnotes (56. 57), both inside a table, but one has its
text inside the cell (56), the other (57) normally with all other notes
at document end. (the former is in a cell with style `a`, the latter
with style `d`).

(See #21, #24)